Added character rotation, jumping, and blocking.
Added the lookAt functionality for the camera, but I don't know if that is needed. I have the camera looking at the player, but the camera is a child of the player. I could programatically set the distance and angle and constantly update the position of the camera within the script as well.